Wireshark: revise registering and unregistering taps (#144)
Refactor how taps are specified, to reduce copy-paste boilerplate.

Check for success when registering each tap, log failures.
(For an example of how this is helpful, this will show that the name of
the TLS tap is no longer "tls" as of the upcoming Wireshark 4.4.)

`init_globals()` gets called every time a file is loaded or a preference
is changed, so add code to `cleanup_globals()` to unregister any taps
registered in `init_globals()`. Avoids leaking unbounded tap
registrations.

Avoid registering taps entirely if the JA4 post-dissector has been
disabled, to improve performance.