Fix SSL version fields: SSL 2.0 is 0x0002, SSL 1.0 never existed
Fix version for SSL 2.0 and remove version for SSL 1.0.

SSL 2.0 [1][2] uses a version field of 0x0002, not 0x0100.  SSL 1 never
existed outside of Netscape, as the original design was iterated upon to
become SSL 2 before the first public version of SSL.  I don't think it's
public knowledge what the version field for SSL 1.0 looked like, or if
it even was two bytes large or at the same offset on the wire.

Version field 0x0100, that JA4 is currently misattributing to SSL 1.0,
was used by an early pre-RFC4347 implementation of DTLS in OpenSSL
before 0.9.8f [2], when OpenSSL switched to the version field specified
by RFC4347.  Arguable if it makes sense to support this early form of
experimental DTLS, I chose not to include it in this PR.

This fix may or may not be related to the TODO comment in the zeek code
that mentions a bug where zeek returns invalid versions; I could not
find the mentioned pcap in the repo so did not investigate further.

[1] https://www-archive.mozilla.org/projects/security/pki/nss/ssl/draft02.html
[2] https://datatracker.ietf.org/doc/html/draft-hickman-netscape-ssl-00
[3] https://github.com/openssl/openssl/compare/OpenSSL_0_9_8e...OpenSSL_0_9_8f