Wireshark crash fixes for 4.4
Fix a crash when fetching `ssh.direction` caused by using the wrong
function in #128. `fvalue_get_uinteger()` doesn't support the
`FT_BOOLEAN` field, but `fvalue_get_uinteger64()` does.

Fix a crash in `locate_tree()` when the first child of the passed-in
tree is non-null but has a null `finfo` member.

These fixes allow the plugin to pass the test suite on the current main
branch (`master`) of Wireshark when built as an in-tree plugin.