Wireshark: correct wmem_array creation
The second argument to `wmem_array_new()` is the size of each element in
the array. The hardcoded value of 100 used in `packet-ja4.c` can cause
runtime errors on some platforms about its values not being properly
aligned in memory to the platform's word size.

Assuming that 100 was meant to be the number of elements to
pre-allocate in the new array, replace `wmem_array_new()` with
`wmem_array_sized_new()` with 100 as the third argument (and a `sizeof`
the element as the second).

Also, in `dissect_ja4()`, replace `wmem_packet_scope()` with
`pinfo->pool`. The two are effectively the same, but this protects the
plugin against [planned improvements][1] to Wireshark memory management.

Built and tested as an in-tree build on Windows 11 against Wireshark's
`master` branch. Ran the test framework from #222 before and after
making my changes, to validate that the output does not change.

[1]: https://lists.wireshark.org/archives/wireshark-dev/202107/msg00052.html